server creation

+++ src/main/java/lk/ijse/gdse/pos/pos_server_javaEE/controller/servlet/ItemServlet.java
@@ -48,13 +48,27 @@ public class ItemServlet extends HttpServlet {
 
         Jsonb jsonb = JsonbBuilder.create();
         ItemDTO itemDTO = jsonb.fromJson(req.getReader(),ItemDTO.class);
-        try {
-            connection = dbcp.getConnection();
-            boolean isSaved = itemBO.saveItem(itemDTO,connection);
-        } catch (SQLException throwables) {
-            throwables.printStackTrace();
-        } catch (ClassNotFoundException e) {
-            e.printStackTrace();
+
+        if (itemDTO.getCode()==null||itemDTO.getCode().matches("/^(I00-)[0-9]{3}$/") ){
+            resp.sendError(HttpServletResponse.SC_BAD_REQUEST,"Code is empty or invalid");
+        }else if (itemDTO.getDescription()==null||itemDTO.getDescription().matches("/^[A-Za-z ]{2,}$/") ){
+            resp.sendError(HttpServletResponse.SC_BAD_REQUEST,"Description is empty or invalid");
+        }else {
+            try {
+                connection = dbcp.getConnection();
+                boolean isSaved = itemBO.saveItem(itemDTO, connection);
+
+                if (isSaved){
+                    resp.setStatus(HttpServletResponse.SC_CREATED);
+                }
+                else {
+                    resp.sendError(HttpServletResponse.SC_INTERNAL_SERVER_ERROR,"Failed to save the item");
+                }
+            } catch (SQLException throwables) {
+                throwables.printStackTrace();
+            } catch (ClassNotFoundException e) {
+                e.printStackTrace();
+            }
         }
 
     }
@@ -68,11 +82,25 @@ public class ItemServlet extends HttpServlet {
 
         Jsonb jsonb = JsonbBuilder.create();
         ItemDTO itemDTO = jsonb.fromJson(req.getReader(),ItemDTO.class);
-        try {
-            connection = dbcp.getConnection();
-            boolean isSaved = itemBO.updateItem(itemDTO,connection);
-        } catch (SQLException | ClassNotFoundException throwables) {
-            throwables.printStackTrace();
+
+        if (itemDTO.getCode()==null||itemDTO.getCode().matches("/^(I00-)[0-9]{3}$/") ){
+            resp.sendError(HttpServletResponse.SC_BAD_REQUEST,"Code is empty or invalid");
+        }else if (itemDTO.getDescription()==null||itemDTO.getDescription().matches("/^[A-Za-z ]{2,}$/") ){
+            resp.sendError(HttpServletResponse.SC_BAD_REQUEST,"Description is empty or invalid");
+        }else {
+            try {
+                connection = dbcp.getConnection();
+                boolean isUpdated = itemBO.updateItem(itemDTO, connection);
+
+                if (isUpdated){
+                    resp.setStatus(HttpServletResponse.SC_CREATED);
+                }
+                else {
+                    resp.sendError(HttpServletResponse.SC_INTERNAL_SERVER_ERROR,"Failed to update the item");
+                }
+            } catch (SQLException | ClassNotFoundException throwables) {
+                throwables.printStackTrace();
+            }
         }
     }
 
@@ -84,11 +112,22 @@ public class ItemServlet extends HttpServlet {
         BasicDataSource dbcp = (BasicDataSource) servletContext.getAttribute("dbcp");
         Connection connection=null;
 
-        try {
-            connection = dbcp.getConnection();
-            boolean isSaved = itemBO.deleteItem(code,connection);
-        } catch (SQLException | ClassNotFoundException throwables) {
-            throwables.printStackTrace();
+        if (code==null||code.matches("/^(I00-)[0-9]{3}$/") ){
+            resp.sendError(HttpServletResponse.SC_BAD_REQUEST,"Code is empty or invalid");
+        }else {
+            try {
+                connection = dbcp.getConnection();
+                boolean isDeleted = itemBO.deleteItem(code, connection);
+
+                if (isDeleted){
+                    resp.setStatus(HttpServletResponse.SC_CREATED);
+                }
+                else {
+                    resp.sendError(HttpServletResponse.SC_INTERNAL_SERVER_ERROR,"Failed to delete the item");
+                }
+            } catch (SQLException | ClassNotFoundException throwables) {
+                throwables.printStackTrace();
+            }
         }
     }
 }
